+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerDefaultDateTimeWithSeconds.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTemplate';
+
+const DateTimePickerDefaultDateTime = () => (
+  <DateTimePickerExampleTemplate
+    value="2017-08-08T10:30:00"
+    showSeconds
+  />
+);
+
+export default DateTimePickerDefaultDateTime;